IMPALA-4239: fix buffer pool test failures in release build
Gtest's ASSERT_DEBUG_DEATH macro has peculiar semantics where in debug
builds it executes the code in a forked process, so it has no visible
side-effects, but in release builds it executes the code as normal. This
makes it difficult to write death tests that work in both debug and
release builds. To avoid this problem, update our wrapper macro to omit
the code in release builds (where we can't actually test DCHECKs
anyway).

Change-Id: Ia560e702ecac2d29dc72f444645d5a91743c95e3
Reviewed-on: http://gerrit.cloudera.org:8080/4596
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins