IMPALA-4204: Remove KuduScanNodeTest
In IMPALA-4204, the limit test could fail depending on the
timing of the scanner threads.

Now that we have functional tests for Kudu (IMPALA-3718), we
can remove the BE KuduScanNode tests which have been error
prone and haven't been any new bugs. Also, we don't have BE
tests for any other scan nodes.

Change-Id: I17c050f461c5cf0649857b82d747d87df0a641b8
Reviewed-on: http://gerrit.cloudera.org:8080/4609
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Internal Jenkins