IMPALA-4291: Reduce LLVM module's preparation time
Previously, when creating a LlvmCodeGen object, we
run an O(mn) algorithm to map the IRFunction::Type
to the actual LLVM::Function object in the module.
m is the size of IRFunction::Type enum and n is
the total number of functions in the module. This
is a waste of time if we only use few functions
from the module.

This change reduces the preparation time of a simple
query from 23ms to 10ms.

select count(*) from tpch100_parquet.lineitem where l_orderkey > 20;

Change-Id: I64e8d1cd00338284c0220cf5458986b73e3d91c2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/21234
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Michael Ho <kwho@cloudera.com>