IMPALA-3420: use gold by default
Also pass the flag that enables ld.gold directly to the
compiler. This is understood by both gcc and clang
(if prefixed with -Wl, clang just forwards the flag to ld,
where it is ignored).

Testing:
Did ASAN and debug private builds to validate it works.
Tested shared library, release, ninja and distcc builds locally
as part of my normal workflow.

Change-Id: Ib05c944ced9cdfe54941f4b690574e45a25110a2
Reviewed-on: http://gerrit.cloudera.org:8080/4751
Reviewed-by: Jim Apple <jbapple@cloudera.com>
Tested-by: Internal Jenkins