IMPALA-4153: Fix count(*) on all blank('') columns - test
This change adds test coverage for the fixes committed for
IMPALA-2399 in commit 9ed3b685a109559fad8dd94f6c03af93e164305a.
It uses the table nulltable in the workload functional-query
to verify the materialization and counting of NULL and empty-
valued columns. The test can be run on any supported storage
and compression combination.

Change-Id: I23923f95f43d67977ee1520a1fc09ce297548b3f
Reviewed-on: http://gerrit.cloudera.org:8080/4755
Tested-by: Internal Jenkins
Reviewed-by: Jim Apple <jbapple@cloudera.com>