IMPALA-4340: explain how to install postgresql-9.5 or higher
The random query generator needs to compare against PostgresQL 9.5 or
higher to take advantage of some of the more recent features, especially
as it pertains to Impala/Kudu INSERT and UPSERT queries.  Developers
will need assistance setting up their development environments if they
need to use the random query generator.

This patch provides instructions on how to do so. We provide
instructions, not automation, since this will have side-effects on
developers' workstations: we can't presume to know how a developer might
want to install or configure postgres, and we haven't tested on anything
except our own development environment.

Change-Id: I1e3b510120451fcb5af97145fa47ccb4c53f00d9
Reviewed-on: http://gerrit.cloudera.org:8080/4846
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>