IMPALA-4388: Fix query option reset in tests
Before this change, using sync_ddl=1 could prevent query options from
being reset correctly. The test execution would use a connection to a
random impalad and execute the test against it, but then undo all
changes to the query options on the default connection (instead of the
one used for the test).

The fix is to undo the changes on the correct connection.

Change-Id: I82e97438ee9f4f75907704653faa884722213f5d
Reviewed-on: http://gerrit.cloudera.org:8080/4870
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins