IMPALA-4411: Kudu inserts violate lock ordering and could deadlock
This fixes a potential hang because the code took
QueryExecState::lock_ before SessionState::lock, which is
the wrong order. A hang has not yet been observed.

The code that was taking the session lock doesn't actually
need to be holding the QueryExecState lock_, so this is easy
to fix by moving the relevant code.

Testing: Running an exhaustive test job, and stress tests
manually.

Change-Id: I2aa36fce78525a80a6b880e1b668105006bc1425
Reviewed-on: http://gerrit.cloudera.org:8080/4895
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins