IMPALA-4412: Per operator timing in profile summary is incorrect when mt_dop > 0
With mt_dop > 0, we may incorrectly populate the
per_fragment_instance_idx for each FInstanceExecParams in
SimpleScheduler::CreateScanInstances().

This causes the averages in the exec summary to be wrong resulting in
misleading Runtime Profiles.

This patch makes sure that each FInstanceExecParams has the correct
instance index associated with it, thereby fixing the exec summaries
in the runtime profiles.

Change-Id: Id8634acba05698281e698ab8ddf92db4ae38af88
Reviewed-on: http://gerrit.cloudera.org:8080/4933
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Internal Jenkins