IMPALA-4423: Correct but conservative implementation of Subquery.equals().
The underlying problem was for trivial/constant [NOT] EXISTS subqueries
we substituted out Subqueries with bool literals using an ExprSubstitutionMap,
but the Subquery.equals() function was not implemented properly, so we ended
up matching Subqueries to the wrong entry in the ExprSubstitutionMap.
This could ultimately lead to wrong plans and results.

Testing: Corrected an existing test and modified an existing test for
extra coverage.

Reviewed-on: http://gerrit.cloudera.org:8080/4923
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 66e9d88c37fc8181dff85f147b22b242869b2cf6)

Change-Id: I61c273a4cc49706ec77058eca48ac58a8bbca296
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/17464
Tested-by: Gaal Laszlo <laszlo.gaal@cloudera.com>
Reviewed-by: Gaal Laszlo <laszlo.gaal@cloudera.com>