IMPALA-4423: Correct but conservative implementation of Subquery.equals().
The underlying problem was for trivial/constant [NOT] EXISTS subqueries
we substituted out Subqueries with bool literals using an ExprSubstitutionMap,
but the Subquery.equals() function was not implemented properly, so we ended
up matching Subqueries to the wrong entry in the ExprSubstitutionMap.
This could ultimately lead to wrong plans and results.

Testing: Corrected an existing test and modified an existing test for
extra coverage.

Change-Id: I5562d98ce36507aa5e253323e184fd42b54f27ed
Reviewed-on: http://gerrit.cloudera.org:8080/4923
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit c5f49ec9bbc1b191545b6edd2848acb632b45973)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/19645
Reviewed-by: Gaal Laszlo <laszlo.gaal@cloudera.com>
Tested-by: Gaal Laszlo <laszlo.gaal@cloudera.com>