IMPALA-4421: Send custom cluster & process failure test results to logs/
Improperly nested quotes in tests/run-custom-cluster-tests.sh caused a
directory named "${RESULTS_DIR}" to be created in the tests directory,
which in turned interfered with run-tests.py.

Additonally noticed that run-process-failure-tests.sh was creating a
results directory within tests/. We shoud not be polluting the tests/
dir with random test artifacts, so changed the result directory path
for those tests as well.

Tested by running both scripts, and ensuring that results wound up in
logs/, and nothing was created in tests/.

Change-Id: I475f61b4ba8a693324bbefed8819d029674b66dd
Reviewed-on: http://gerrit.cloudera.org:8080/4918
Reviewed-by: Jim Apple <jbapple@cloudera.com>
Tested-by: Internal Jenkins