IMPALA-4391: fix dropped statuses in scanners
As far as I'm aware we haven't seen any failures related to these in
practice, but fixing them as a preventative measure.

Testing:
I was unable to reproduce this easily by adding a failpoint. I suspect
that stress testing of the affected file formats would have eventually
found this, because of the similarity to IMPALA-3962.

Change-Id: I7c47f8b29a20dc74ad9e9e1c58b3d7ed95de4d35
Reviewed-on: http://gerrit.cloudera.org:8080/4938
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins