IMPALA-4437: hit DCHECK in buffered-block-mgr-test
The issue was introduced by "IMPALA-3202: DiskIoMgr improvements
for new buffer pool". The bug is pretty straightforward - I forgot
to update some error/cancellation paths to handle client-provided
buffers.

Testing:
Was able to reproduce locally before the fix, have run it in a loop for
a while with the fix and not seen any crashes.

Change-Id: Ib7e8ffdbec341f6ffbc310fb29df63ab45fed545
Reviewed-on: http://gerrit.cloudera.org:8080/4973
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins