IMPALA-4494,IMPALA-4540: Fix crash in SimpleScheduler
The scheduler maintains a local list of active backends, which is
updated through messages from the statestore. Even the local backend
enters this list by registering with the statestore and being included
in a statestore update message. Thus, during restarts it can happen that
a query gets scheduled with exec_at_coord set to true, while the local
backend has not been registered with the scheduler. In this case the IP
address lookup in the internal BackendConfig fails and an empty IP
address is returned, leading to a nullptr dereference down the line.

This change adds an additional check when handling updates from the
statestore to make sure that the backend config always contains the
local backend. It also changes scheduling when exec_at_coord is true to
always use the local backend, irrespective of whether it is present in
the backend config.

Change-Id: I6e1196a2fa47e5954c4a190aa326c135d039a77f
Reviewed-on: http://gerrit.cloudera.org:8080/5127
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 96d98abff52e59742469f9d5a86018506e00f88f)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/21043
Tested-by: Attila Jeges <attilaj@cloudera.com>
Reviewed-by: Lars Volker <lv@cloudera.com>