IMPALA-4512: Add a script that builds Impala on stock Ubuntu 14.04.
This is a simpler alternative to bootstrap_development.sh - it
acquires enough dependencies to build, but does not attempt to load
the test data or even build the tests. This is sometimes a lightweight
testing method used by Apache PPMC members who are voting on a release
of an incubating project.

Change-Id: If34e398052a61dfda9825b1cf3a918eb61736048
Reviewed-on: http://gerrit.cloudera.org:8080/5154
Reviewed-by: Jim Apple <jbapple-impala@apache.org>
Tested-by: Impala Public Jenkins