IMPALA-4522: Bound Kudu client threads to avoid stress crash
In stress testing on physical boxes (80 cores, 200gb ram) we
discovered that the Kudu Java client creates a huge number
of threads (2x the #cores) per Kudu client, and this was
causing the impalad to crash when the JVM couldn't create
more threads.

This addresses the issue by setting the number of Kudu
client worker threads rather than letting the Kudu client
pick the default (2 * #cores). The number set here was
suggested by the Kudu team as being sufficient for Impala's
FE usage and this has been tested for 8+ hours on the stress
cluster where the crash was previously observed quickly.

In the future, Impala should probably be sharing a single
Kudu client (it is multithreaded), but additional support
from Kudu may be needed to ensure this usage is correct
(e.g. client metadata may need invalidation after some
operations).

Change-Id: I3940df776eaa5ad22e1bbb572559afcc8990bf1d
Reviewed-on: http://gerrit.cloudera.org:8080/5205
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins