IMPALA-4550: Fix CastExpr analysis for substituted slots
During slot substitution, the type of the child of a CastExpr can
change. If the previous child type matched the CastExpr, then the cast
was flagged as noOp_. During substitution and subsequent re-analysis
the noOp_ flag was not revisited so that no cast was performed, even
after it had become necessary.

The fix is to always set noOp_ to the correct value in
CastExpr.analyze().

Change-Id: I7f29cdc359558fad6df455b8eec0e0eaed00e996
Reviewed-on: http://gerrit.cloudera.org:8080/5267
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 0f62bf35fd2266d130cf06de0710b7d7797d84ac)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/19646
Reviewed-by: Gaal Laszlo <laszlo.gaal@cloudera.com>
Tested-by: Gaal Laszlo <laszlo.gaal@cloudera.com>