IMPALA-4577: Adjust maximum size of row batch queue with MT_DOP.
When MT_DOP is set, non-Parquet scans are run with the old
scan node with a fixed NUM_SCANNER_THREADS=1. This patch
adjust the maximum size of the row batch queue based on MT_DOP
for each such scan instance to avoid a significant increase in
the memory consumption of such scans. The max queued batches
per scan-node instance is at least 2 to always allow for some
parallelism between the producer/consumer.

Decreases the maximum allowed value for MT_DOP from 128 to 64.

Change-Id: Ic2aa260f9265ec21173fb703c41934964ece6485
Reviewed-on: http://gerrit.cloudera.org:8080/5330
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins