IMPALA-4608: Fix fragment completion times for INSERTs
Fix a bug where completion times weren't computed if the query was an
INSERT, because the code presumed that instance 0 was always the
coordinator fragment and skipped completion time computation.

It may be that the special-casing can be removed entirely, but needs
further investigation to make sure that wouldn't trigger any div-by-0 bugs.

Change-Id: I3ce56f70d30c9e398b14b32520c815d87f81f893
Reviewed-on: http://gerrit.cloudera.org:8080/5418
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Impala Public Jenkins