IMPALA-3202: variable-length scratch file ranges
This uses a simple approach where scratch ranges are managed in
power-of-two size classes and we don't attempt to coalesce or
split the ranges to move them into different size classes. Thus
this does not optimally reuse space if a query spills a variety
of page sizes, but improving this may not be worth the added
complexity.

Testing:
Extended tmp-file-mgr-test to exercise the variable scratch range
support. We will get system test coverage once this is used by the
new buffer pool.

Change-Id: Ic0ad84493c2c93a5602c404a83c718f25ea25575
Reviewed-on: http://gerrit.cloudera.org:8080/5597
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins