CDH-48554: Fix refresh of empty partition directories.
Running REFRESH on deleted table/partition directories does not
update the internal file metadata listings of the partitions
corresponding to these paths. This can result in an inconsistent
state of the Catalog and can potentially fail queries since they
try to scan non-existent paths.

Fix: If we encounter an empty partition directory, we need to
clean up the existing partition state in the Catalog which includes
the list of fds, no. of files etc., before we proceed further.

Workaround: Use "invalidate metadata" or "alter table recover
partitions"

Note: This only applies to 5.7.x, 5.8.x and 5.9.x. This is fixed
in 2.8/5.10.0 via IMPALA-4172 and hence it doesn't apply to master
anymore.

Change-Id: I160ac34f0bb9be7818d52a4d9d95d9d6a1893a13