IMPALA-3402: Change version numbers in shared variables file.
Some instances of 'CDH' in the doc metadata might show up
in variable substitions or unexpected places generally,
such as the PDF title page or inside cover. Switching them
all to Impala-only numbers. Filling in Impala 2.8 as the
'current' number to reflect what will be the next release.

Also remove another instance of 'Cloudera' in the title.

Change-Id: I69e855aa958a45b3286a27fa69bcf8df315fc174
Reviewed-on: http://gerrit.cloudera.org:8080/5626
Reviewed-by: Jim Apple <jbapple-impala@apache.org>
Tested-by: John Russell <jrussell@cloudera.com>