IMPALA-4716: Expr rewrite causes IllegalStateException
The DECODE constructor in CaseExpr uses the same decodeExpr object when
building the BinaryPredicates that compare the decodeExpr to each 'when'
of the DECODE. This causes problems when different BinaryPredicates try
to cast the same decodeExpr object to different types during analysis,
in this case leading to a Precondition check failure.

The solution is to clone the decodeExpr in the DECODE constructor in
CaseExpr for each generated BinaryPredicate.

Testing:
- Added a regression test to exprs.test

Change-Id: I4de9ed7118c8d18ec3f02ff74c9cca211c716e51
Reviewed-on: http://gerrit.cloudera.org:8080/5631
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 89a3d3c1ebf85629bb20457a9672d71fbf2504ca)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/23246
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>
Reviewed-by: Attila Jeges <attilaj@cloudera.com>
(cherry picked from commit 0d7674d744250fefeb59df987edaa616a8cfb355)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/24608
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>