Patch references to Cloudera and CDH in Impala tutorial
There was one tutorial that actually ran under the
'cloudera' user and so repeated that name over and
over in directory and HDFS paths. I switched that to
'username'.

I suppressed some <note> and <li> tags with Cloudera
Manager-specific details. Will physically remove those
from the source in a subsequent iteration.

I left several instances of audience="Cloudera" because
those will be changed to audience="hidden" as part of
a separate change request.

I marked with rev="upstream" some <codeblock> tags
containing impala-shell banners with a Cloudera copyright
statement. Will decide on a convention to handle those
(elide those lines, or use a conref to consistently
substitute the generic equivalent) and do that in a
followup patch set.

Change-Id: I44245b65ce6f247ae8771f582f4b33c3712145ae
Reviewed-on: http://gerrit.cloudera.org:8080/5663
Reviewed-by: Laurel Hale <laurel@cloudera.com>
Reviewed-by: Jim Apple <jbapple-impala@apache.org>
Tested-by: Impala Public Jenkins