IMPALA-3748: add query-wide resource acquisition step
This adds a Prepare() method to QueryState that allows
query execution to fail up-front if resources cannot be
acquired. We don't make full use of this yet, but it
provides a suitable place to acquire a memory reservation
for the query or fail cleanly.

The process memory limit check is moved here so that all
startup memory checks will be eventually consolidated in
one place.

Also switch a boost::mutex to a SpinLock for consistency
and to improve performance (see lock-benchmark.cc).

Change-Id: Ia21a3c0f0b0a7175116883ef9871b93c8ce8bb81
Reviewed-on: http://gerrit.cloudera.org:8080/5739
Tested-by: Impala Public Jenkins
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>