Improve substitution variables.
Fix the version numbers that show up in titles
and in lead-ins, e.g. "____ or higher only", to
reflect purely Impala release numbers.

Substitute the most prevalent forms of reference
to CDH and Cloudera support with generic equivalents.

Resolved merge conflict with the code review to reorg
all the keydefs into a separate file.

Change-Id: Ia23b96a8344e6430528701b81bea7eaaccc72b18
Reviewed-on: http://gerrit.cloudera.org:8080/5661
Reviewed-by: John Russell <jrussell@cloudera.com>
Tested-by: Impala Public Jenkins