IMPALA-4801: fix heap use after free for MemTracker
This is a minimal fix for IMPALA-4801. The bug was that
"IMPALA-4678: move query MemTracker into QueryState" changed
the lifecycle of RuntimeState::instance_mem_tracker_ so that
it destroyed earlier in ReleaseResources(). A periodic
counter referencing the MemTracker was only stopped after
ReleaseResources() was called, leaving a window for a race.

There are various issues with the lifecycle of RuntimeProfile
objects (see IMPALA-4802) but I chose not to tackle them in
this patch.

Change-Id: Iccaf6700888c333b9c93d716d395d96bc5af6631
Reviewed-on: http://gerrit.cloudera.org:8080/5772
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins