IMPALA-3406: [DOCS] Empty the original Cloudera FAQ
Almost all of the original Impala FAQ material was
Cloudera-themed or commercially oriented. Lots of
answers about the QuickStart VM, Cloudera discussion
forums, CDH-based recommendations, etc. IMO it is
not worth trying to adapt each FAQ entry to be generic.
Better to start over from the ground up.

Phase 1 of making an Apache-friendly FAQ is to strip
the original page "down to the studs" so new FAQ
entries can be added with more of a developer theme,
based on questions people have in the community.

Change-Id: Ib81242f0981c04fff99e2c27e06a8d9f4da34c9f
Reviewed-on: http://gerrit.cloudera.org:8080/6003
Reviewed-by: Jim Apple <jbapple-impala@apache.org>
Tested-by: Impala Public Jenkins