IMPALA-3401 [DOCS] Part2 Removing 'Cloudera Manager' from upstream docs.
Used DITA attribute 'audience=hidden' when needed to preserve
information for a future clean-up step. When it made sense,
'Cloudera Manager' was removed and sections were rewritten.
This clean-up task does not completely remove 'Cloudera Manager'
from the upstream docs source XML, but it is now
hidden in these files so it won't show up in the rendered
version. A subsequent cleanup project will remove references
in the XML source.

Change-Id: I76c9b53f587bc85c5c21e195f0a771183d4ef3a0
Reviewed-on: http://gerrit.cloudera.org:8080/6064
Reviewed-by: John Russell <jrussell@cloudera.com>
Tested-by: Impala Public Jenkins