IMPALA-3401 [DOCS] Phase 5 of "Cloudera Manager" removal.
Most of these fixes entailed hiding the
paragraph where there are mentions of CM and
writing a replacement paragraph for the upstream
docs. This removes the CM references from the
rendered docs. A subsequent cleanup project will
remove occurrences of CM from the XML. This patch
includes fixes to files in response to John
Russell's first set of review comments.

Change-Id: I4967fae275a8822274aece14a15107237445aba5
Reviewed-on: http://gerrit.cloudera.org:8080/6070
Reviewed-by: John Russell <jrussell@cloudera.com>
Tested-by: Impala Public Jenkins