IMPALA-3401 [DOCS] Phase 7 of "Cloudera Manager" removal
Removed the topic "impala_howto_rm.xml" from the build
because it was focused on resource management by using
Cloudera Manager. Commented it out of impala.ditamap,
impala_keydefs.ditamap, and from a reference to it
in shared/impala_common.xml. This removes it
completely from the rendered upstream docs. A later
project will remove the source XML.

Change-Id: I26fc661144ba906828d5f6b5b2ea2eca02693369
Reviewed-on: http://gerrit.cloudera.org:8080/6108
Reviewed-by: John Russell <jrussell@cloudera.com>
Tested-by: Impala Public Jenkins