IMPALA-4997: Fix overflows in Sorter::TupleIterator
Various places in Sorter::TupleIterator multiply two int values
(Sorter::Run::block_capacity_ and Sorter::TupleIterator::block_index_)
and assigned the result to an int64_t value
(Sorter::TupleIterator::buffer_start_index_). One such occurrence is in
be/src/runtime/sorter.cc#L1080. This multiplication could overflow for
runs with a large number of rows. Changing one of the operands to
int64_t fixes this.

To test this Matt Mulder ran the failing query from IMPALA-4997 on a
private cluster and it succeeded.

Change-Id: Iea22aa96e0cc86102b60c6e551e9e607cef485c8
Reviewed-on: http://gerrit.cloudera.org:8080/6169
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins