IMPALA-5027: make udf headers buildable externally
This fixes a number of issues that prevented building it:
* Use of C++11 features
* Use of internal headers that are only available in the Impala source
  tree

Testing:
Copied the headers and libImpalaUdf.a to /usr and confirmed I could
build impala-udf-samples (with some modifications to udf-test-harness
to use some modified arguments to the test harness).

We really need a better solution for automatically testing this, e.g.
IMPALA-5026, but I wanted to fix the known bugs first.

Change-Id: Ibdcdcd96c953cbabc3de04ca66d7aa1774ada99e
Reviewed-on: http://gerrit.cloudera.org:8080/6251
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Impala Public Jenkins
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/20884
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>