IMPALA-5038: Fix file size regex to include bytes
There is a regex to remove file sizes from test logs to avoid diffs
caused by changes in file headers, etc. The regex was detecting
size=number correctly, but it does not include the byte unit (i.e.
size=900B is distinct from size=900KB). A file size change introduced
by IMPALA-4624 caused a diff because the file size changed from 900B
to 1.1KB. This fixes the regex to include the byte unit, so that
changes of this form do not cause diffs.

Change-Id: I5f7b8480065a4da63c8abaf53aae8aef772f0172
Reviewed-on: http://gerrit.cloudera.org:8080/6288
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Impala Public Jenkins