IMPALA-4539: fix bug when scratch batch references I/O buffers
The bug occurs when scanning uncompressed plain-encoded Parquet
string data.

Testing:
I could manually reproduce it reliably with ASAN and
--disable_mem_pools=true using the following steps:

  # The repro is more predictable when there is no row batch queue.
  set mt_dop=1;
  # A unique string column should switch to plain encoding
  set compression_codec=none;
  create table big_uuids stored as parquet as select uuid() from tpch_20_parquet.lineitem;
  # The repro requires that some rows are filtered out, so that we end
  # up in a state where the output batch is full before all rows are
  # copied from the scratch batch
  select ndv(_c0) from big_uuids where substring(_c0, 1, 2) != 'a1' limit 10;

After the fix it no longer reproduces.

I do not yet have a practical test case that triggers the bug on a
normal ASAN setup. I will continue to try to create one.

NOTE FOR CDH 5.8.x
This is an equivalent re-implementation of the fix for IMPALA-4539 because
of significant code differences between CDH 5.8.x and the original branch
where the issue was fixed (CDH 5.10.0).

Change-Id: Ia136c7684eab9ff2d8bf99a69e9113dbc13933d2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/20923
Reviewed-by: Alexander Behm <alex.behm@cloudera.com>
Tested-by: Alexander Behm <alex.behm@cloudera.com>