IMPALA-5055: Fix DCHECK in parquet-column-readers.cc ReadPageHeader()
GetBytes only sets status in the case of an error. This means that
ReadPageHeader needs to initialize the status variable so that the
status.ok() check is accurate after the GetBytes call.

I verified that the other uses of status are ok. Most do not check
status.ok() directly, but rely on the return value of the function
setting status.

Change-Id: Ie22a8cf6b53f507c378c2efe302482409935184e
Reviewed-on: http://gerrit.cloudera.org:8080/6328
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Impala Public Jenkins