IMPALA-5055: Fix DCHECK in parquet-column-readers.cc ReadPageHeader()
GetBytes only sets status in the case of an error. This means that
ReadPageHeader needs to initialize the status variable so that the
status.ok() check is accurate after the GetBytes call.

I verified that the other uses of status are ok. Most do not check
status.ok() directly, but rely on the return value of the function
setting status.

Notes for CDH5.11.x: The upstream original file diverged quite a bit
but the relevant change is the same: status needs to be initialized
in exactly the same way as in the upstream code.

Change-Id: Ie22a8cf6b53f507c378c2efe302482409935184e
Reviewed-on: http://gerrit.cloudera.org:8080/6328
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 244ec22d8e8b2f108f723d49dc95a016a74d6d6f)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/22615
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>