[DOCS] Genericize CREATE FUNCTION page
Several instances of things to revise on this page.
For example, com/cloudera in paths, -cdh570 in
filenames, and cloudera in github URLs.

There will still be some cloudera github URLs in the
HTML output after this change, but the source of the
CREATE FUNCTION page itself will be clean. (The URLs
in question will be edited in impala_keydefs.ditamap
as part of the endgame to revise various github URLs
related to UDF development.

Change-Id: I42fc2c59e744e43504ad92dc94f6659a59f49bd2
Reviewed-on: http://gerrit.cloudera.org:8080/6371
Reviewed-by: Ambreen Kazi <ambreen.kazi@cloudera.com>
Reviewed-by: John Russell <jrussell@cloudera.com>
Tested-by: Impala Public Jenkins