CDH-51687: Readjust timestamp MAX_YEAR to 10000
This patch compensates for the lack of IMPALA-4549 in the commit
chain of 5.10.1 and 5.11.0. With that fix missing, the maximum
year value of a timestamp should be allowed to reach 10000
(IMPALA-4549 would limit that to 9999).

The patch also reinstates a test step in expr-test.cc checking
that in a date calculation the year part of the intermediate result
can reach the maximum value of 10000.

Change-Id: I39aa39b00f989d29d7760077dda3f73f6ee02b17
(cherry picked from commit 54ffd77b70dbff254d7393da1a7490d0a3c24326)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/21248
Reviewed-by: Michael Ho <kwho@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>