IMPALA-5088: Fix heap buffer overflow
In a recent patch (IMPALA-4787), we introduced a change where we are
reading past the end of a buffer during a copy. Even though we would
never read the data that was copied from past the end of the buffer,
this could cause a crash if the allocation is sitting on the edge of an
unmapped area of memory. This also caused the ASAN build to fail.

The issue is fixed by never accessing memory that is past the end of the
buffer.

Change-Id: Ia68d69b35d53934e6f1aae2e1696503d17660ca2
Reviewed-on: http://gerrit.cloudera.org:8080/6441
Reviewed-by: Taras Bobrovytsky <tbobrovytsky@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit c27ad3526675ff0ab80c7bb5d7c8364437448b53)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/22430
Tested-by: Attila Jeges <attilaj@cloudera.com>
(cherry picked from commit f725871bf30c5e6b257c70ecf3b1e38b315f1348)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/23327
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>