IMPALA-5123: Fix ASAN use after free in timezone_db
The issue is that the string temporary returned by .string goes
out of scope immediately after being created.  Also, the API
to mkstemp is unclear on whether it modifies the string in place.
Just strdup() the c_str() to be safe - this is not performance
critical code.

Testing: ASAN build, running expr-test be test; ASAN fails before,
passes after this change.

Change-Id: I490f741403ea2004bc51394aa1251577337b1e1d
Reviewed-on: http://gerrit.cloudera.org:8080/6503
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins