CDH-52202: Fix setup logic in test_udfs.py
test_udfs.py prepares the test by dropping all UDFs that might have
been declared in a previous run (they are persisted in HMS).
A newly added UDF, to_upper() was missing this setup logic,
which caused test failures when the test was run with a preexisting
metastore; this patch adds to_upper() to the list of UDFs dropped.

Change-Id: I2a3d930d2a541bb7853174283370e524aecf6bdd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/21537
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>