IMPALA-5154: Handle 'unpartitioned' Kudu tables
The catalogd was hanging trying to load an unpartitioned
Kudu table created outside of Impala. This fixes an
assumption made in KuduTable.java that the list of
'partition by' expressions is not empty. Regardless, the
list on the thrift structure must be created because the
field is marked required.

Change-Id: I40926bf6ea46cfca518bba6d4ca13fb5b0de358d
Reviewed-on: http://gerrit.cloudera.org:8080/6560
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 58286bda7a62251f3841fa431013bdb009c6bfaf)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/24674
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Attila Jeges <attilaj@cloudera.com>