IMPALA-5172: Buffer overrun for Snappy decompression
When using a preallocated buffer for decompression, a
file corruption can lead to the expected decompressed size
being smaller than the actual decompressed size. Since
we use this for allocating the output buffer,
decompression needs to be able to handle a buffer that
is too small.

Snappy does not properly handle a buffer that is too small
and will overrun the buffer. This changes the code to
check the decompressed length and return an error if
the buffer is not large enough. It also adds a test to
verify that this behavior is respected for other
compression algorithms.

Change-Id: I45b75f61e8c0ae85f9add5b13ac2b161a803d2ba
Reviewed-on: http://gerrit.cloudera.org:8080/6625
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Dan Hecht <dhecht@cloudera.com>
(cherry picked from commit fcefe47d7348f01e6a6fef700b421290a3f536b3)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/24621
Tested-by: Attila Jeges <attilaj@cloudera.com>
Reviewed-by: Gabor Kaszab <gaborkaszab@cloudera.com>
Reviewed-by: Attila Jeges <attilaj@cloudera.com>