IMPALA-5230: fix non-functional impalad under ASAN
The bug is that TCMalloc metrics are not initialised under
ASAN but other metrics that reference them are created. As
soon as a debug webpage tries to access that metric, Impala
crashes.

The bug was introduced by commit 955b257cfb5: "IMPALA-5073:
Part 1: add option to use mmap() for buffer pool".

Testing:
Started the cluster under ASAN, ran some query tests, checked the /memz
and /metrics pages.

Change-Id: Ic71fa57e0444f470398e16513e4ba6e7f71565e4
Reviewed-on: http://gerrit.cloudera.org:8080/6694
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Impala Public Jenkins