IMPALA-4779, IMPALA-4780: Fix conditional functions built-in and Timestamp bounds
IMPALA-4779: conditional function isfalse(), istrue(),
isnotfalse(), isnotrue() aren't cross-compiled so they
will lead to unexpected query failure when codegen is
enabled. This change will cross-compile these functions.
Some tests are added for this case.

IMPALA-4780: The class static variable TimestampFunctions::MIN_YEAR
isn't a compile time constant. It relies on llvm.global_ctor to be
called to initialize it. However, llvm.global_ctor is never called
in Impalad so that variable is never set, leading to wrong value for
it and other variables dependent on it. This change fixes the problem
above by setting TimestampFunctions::MIN_YEAR to a known constant
value. A DCHECK is added to verify that it matches the value
defined in the boost library. Also fixes the incorrect computation
of TimestampFunctions::MAX_MINUTE_INTERVAL.

Change-Id: I6f904de68d5e0780b783a1867e794c1e84b1f178
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/22157
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: Michael Ho <kwho@cloudera.com>