IMPALA-5291: avoid calling global destructors in statestore-test
The workaround implemented by this patch is to exit the process using
_exit(), which does not run global destructors.

The "proper" solution would be either to change Thrift so that the
destructors are not run unnecessarily on process teardown, or to extend
the statestore services and clients to allow stopping them. It seems the
complexity of neither is justified.

Testing:
Was able to reproduce the bug by running statestore-test in a loop.
After the fix it was not reproducible.

Change-Id: Ic185825cf262311bdc05784ebd541cf53a13cdd6
Reviewed-on: http://gerrit.cloudera.org:8080/6872
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Impala Public Jenkins