IMPALA-4764: Add Hedged read metrics
The following Hadoop metrics have been added to the /metrics page:

hedgedReadOps - the number of hedged reads that have occurred

hedgedReadOpsWin - the number of times the hedged read returned
faster than the original read

The metrics will be updated only when --use_hdfs_pread is set to
'true'.

This change depends on the following new commit to HDFS:
https://github.com/apache/hadoop/commit/8c81a16a1fa40d3a3796b1c3633328f1340883ca

Testing: Not adding tests since it requires some custom hadoop
configuration. Tested manually by setting the configurations and
verifying that the metrics work.

Change-Id: Id4a5d396abb3373d352ad2df8c2272db018114da
Reviewed-on: http://gerrit.cloudera.org:8080/6886
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins