IMPALA-5364: Correct title of query locations table
Previously claimed to be #fragments, but this was inaccurate - it's the
number of queries with fragments on that host.

We should have per-backend reporting on the number of fragment instances
in the future, but for now let's correct the mistaken label.

Change-Id: I2fddd41500ada10eb1c12d92afcb96a82009857b
Reviewed-on: http://gerrit.cloudera.org:8080/6988
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Impala Public Jenkins
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/23498
Tested-by: Henry Robinson <henry@cloudera.com>
(cherry picked from commit 70210e71cc3178132dee2c4f62c479d11807f95f)