IMPALA-5355: Fix the order of Sentry roles and privileges
After a single Impalad is restarted, it is possible that order in which
it receives roles and privileges from the statestore is incorrect. The
correct order is for the role to appear first in the update, before the
privilege that references it.

If a user updates a role, its catalog version number can become larger
than the catalog numbers of the privileges that reference it. This
causes the role to come after the privilege in the initial metastore
update.

The issue is fixed by doing two passes over the catalog objects in the
Impalad. The first pass updates the top level objects. The second pass
updates the dependent objects

Testing:
- Added a test that reproduced the problem.

Change-Id: I7072e95b74952ce5a51ea1b6e2ae3e80fb0940e0
Reviewed-on: http://gerrit.cloudera.org:8080/7004
Reviewed-by: Taras Bobrovytsky <tbobrovytsky@cloudera.com>
Tested-by: Impala Public Jenkins
(cherry picked from commit 6604083f518d332f3f13601e61a9860c23cab4ca)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/24618
Tested-by: Attila Jeges <attilaj@cloudera.com>